refactor anova package
