+++ src/main/java/io/github/mzmine/modules/dataanalysis/significance/anova/help/csv_file.png
