+++ src/main/java/io/github/mzmine/modules/dataanalysis/significance/anova/help/help.html
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright (c) 2004-2022 The MZmine Development Team
+  ~ Copyright (c) 2004-2024 The MZmine Development Team
   ~
   ~ Permission is hereby granted, free of charge, to any person
   ~ obtaining a copy of this software and associated documentation
